Move filter components into separate files.
