+++ tests/unit/modules/election-lookup.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 /* global window */
