+++ tests/unit/modules/filter.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 var chai = require('chai');
 var expect = chai.expect;
